Added parallelization for buddy, isolation, and metadata checks
Note this changes the behaviour of buddy check, in that within an
iteration, the earlier flagging of a station is not taken into account
when computing the neighbourhood statistics of a later station. This
would be impossible to do with parallelization. The other advantage of
the new behaviour is that the order that observations are processed does
not impact the end result.