Change behaviour of 'indices' in Dataset
For methods like SCT and buddy check, where a specific location depends
on other locations, run the test on all non-flagged stations (including
those not specified by indices) but only save the flags for
stations specified by indices. This is different than the previous
behaviour, where the tests were only run on the subset of stations
specified by indices. For isolation test, this isn't desired, because
typically we only want to run this on some stations, but we want all
stations to be part of the isolation count.